ISI-407 - Datentyp geändert, DatabaseFiller erweitert

+++ src/main/java/de/muenchen/isi/api/dto/BedarfsmeldungFachabteilungenDto.java
@@ -10,15 +10,15 @@ import lombok.ToString;
 @EqualsAndHashCode(callSuper = true)
 public class BedarfsmeldungFachabteilungenDto extends BaseEntityDto {
 
-    private Long anzahlEinrichtungen;
+    private Integer anzahlEinrichtungen;
 
     private InfrastruktureinrichtungTyp infrastruktureinrichtungTyp;
 
-    private Long anzahlKinderkrippengruppen;
+    private Integer anzahlKinderkrippengruppen;
 
-    private Long anzahlKindergartengruppen;
+    private Integer anzahlKindergartengruppen;
 
-    private Long anzahlHortgruppen;
+    private Integer anzahlHortgruppen;
 
-    private Long anzahlGrundschulzuege;
+    private Integer anzahlGrundschulzuege;
 }
